Capstone Project
The game difficulty now increases. The grid size has been changed to 2x2, in order to accomodate for future grid resizing based on difficulty, and the sequence length increases +1 once the player gets the sequence correct.

NEXT STEPS: the sequences are obviously easy to remember on a 2x2 grid, after a few levels, let's say 4, the grid will resize to 3x3, and so on, with a simplified sequence to start, and then it will increasingly get harder.